+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file - [read more
 
 ## [Unreleased]
 
+### Added
+- Symfony 2.3 web tests for resource name #349
+
 ### Fixed
 - Resource name on Symfony 2.x requests served through controllers #341
 - Sanitize url in web spans #344